iRIC_Installer_v4beta.exe updated.
Old one was built with Qt installer framework 3.2, but new one is built with Qt Installer framework 2.0
When iRIC v4 was installed with old one, "installerResource" folder was added in each folder, that looks like temporary folders. It looks that it is a bug that temporary files are not removed after installation.

Installer built with Qt Installer framework 2.0 dows not have that issue, so I've replaced iRIC_Installer_v4beta.exe.